src/requirements.txt
mido==1.2.9
python-rtmidi==1.4.7
nanoleafapi==2.0.0
RPi.GPIO
